requirements.txt
fastapi==0.103.2
google-generativeai==0.5.2 
uvicorn==0.23.2
selenium==4.20.0
pydantic==2.4.2